Do not use -f with readlink
Using readlink with -f results in outputting the full path of the
symlink we're referring. Since we are doing a "cut -d'v' -f2", for
extracting version number out of the versions (e.g. 23 out of v23),
we better not use the full pathname to the file, as the cut will
split over a 'v' if it is present in the full address.

It was my mistake -- I thought -f meant 'following symlinks
transitively till how much ever depth it can go'.