Temporarily run acceptance tests every half-an-hour
Note that this is only done temporarily. We will set the tempest test to run every couple of mins or so. As acceptance tests are going to take around an hour to complete (as opposed to a minute it is taking right now), this will make sure our build queue is not empty for extended duration.

How to configure jenkins when we will start parallellizing builds for this stage is not been thought of yet